refactor: quit extracting `isSupportedFieldType()`
